// Angular/src/app/Models/user/Register.User.Model.ts
export class RegisterUserModel {
    Action: string;
    UID_UserId: number;
    UID_CompanyName: string;
    UID_PumpRegNumber: string;
    UID_OilCompany: string;
    UID_Address: string;
    UID_State: string;
    UID_City: string;
    UID_Pin: string;
    UID_UserType: string;
    UID_MobileNumber: string;
    UID_eMail: string;
    UID_GSTIN: string;
    UID_PAN: string;
    UID_TIN: string;
    UID_CIN: string;
    UID_ValidFrom: Date;
    UID_ValidThru: Date;
    ULA_FirstName: string;
    ULA_LastName: string;
}